refactor(tools): extract validateSearchInput from the search_emails handler (Phase 0)
The search_emails handler carried ~110 LOC of arg validation (type/length/range
checks, VALID-002 length caps, TOOL-016 byte predicates, TOOL-017 parseable
dates) interleaved with orchestration — the worst mixed-responsibility spot in
the tool layer per the hotspot review. Move it verbatim into a focused
src/tools/search-input.ts → validateSearchInput(args, maxEmailListResults) that
returns a ready SearchEmailOptions or throws McpError. The handler is now 4
lines. Every check + message preserved (input-hygiene.test is the guard); new
direct unit tests added. Suite 2104 green.
